Added phone number box in register.dart
