+++ lib/DashBoard.dart
@@ -1,32 +1,61 @@
-import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
+import 'package:curved_navigation_bar/curved_navigation_bar.dart';
 
-class DashBoard extends StatelessWidget{
+void main() => runApp(MaterialApp(home: BottomNavBar()));
+
+class BottomNavBar extends StatefulWidget {
+  @override
+  _BottomNavBarState createState() => _BottomNavBarState();
+}
+
+class _BottomNavBarState extends State<BottomNavBar> {
+  int _page = 0;
+  GlobalKey<CurvedNavigationBarState> _bottomNavigationKey = GlobalKey();
+
   @override
   Widget build(BuildContext context) {
     return Scaffold(
-        appBar: AppBar(
-          title: Text('DashBoard'),
+        bottomNavigationBar: CurvedNavigationBar(
+          key: _bottomNavigationKey,
+          index: 0,
+          height: 60.0,
+          items: <Widget>[
+            Icon(Icons.groups_rounded, size: 40),
+            Icon(Icons.group_rounded, size: 40),
+            Icon(Icons.add_circle_sharp, size: 50),
+            Icon(Icons.call_split, size: 40),
+            Icon(Icons.account_circle_rounded, size: 40),
+          ],
+          color: Colors.white,
+          buttonBackgroundColor: Colors.white,
+          backgroundColor: Colors.blueAccent,
+          animationCurve: Curves.easeInOut,
+          animationDuration: Duration(milliseconds: 600),
+          onTap: (index) {
+            setState(() {
+              _page = index;
+            });
+          },
+          letIndexChange: (index) => true,
         ),
-        body: Center(
-          child: Column(
-            mainAxisAlignment: MainAxisAlignment.center,
-            crossAxisAlignment: CrossAxisAlignment.center,
-
-            children: [
-              Text('Welcome', style: TextStyle(
-                fontSize: 34,
-                fontWeight: FontWeight.bold,
-
-              ),),
-              SizedBox(
-                height: 11,
-              ),
-              ElevatedButton(onPressed: (){}, child: Text('Next'))
-            ],
+        body: Container(
+          color: Colors.blueAccent,
+          child: Center(
+            child: Column(
+              mainAxisAlignment: MainAxisAlignment.center,
+              children: <Widget>[
+                Text(_page.toString(), textScaleFactor: 10.0),
+                ElevatedButton(
+                  child: Text('Go To Page of index 1'),
+                  onPressed: () {
+                    final CurvedNavigationBarState? navBarState =
+                        _bottomNavigationKey.currentState;
+                    navBarState?.setPage(1);
+                  },
+                )
+              ],
+            ),
           ),
-        )
-    );
+        ));
   }
-
 }
