+++ lib/register.dart
@@ -1,4 +1,10 @@
 import 'package:flutter/material.dart';
+import 'package:intl_phone_field/intl_phone_field.dart';
+
+
+
+void main() => runApp(MaterialApp(home: MyRegister()));
+
 
 class MyRegister extends StatefulWidget {
   const MyRegister({Key? key}) : super(key: key);
@@ -113,6 +119,45 @@ class _MyRegisterState extends State<MyRegister> {
                           ),
                           SizedBox(
                             height: 40,
+                          ),
+
+
+                          IntlPhoneField(
+                            decoration: InputDecoration(
+                                enabledBorder: OutlineInputBorder(
+                                  borderRadius: BorderRadius.circular(10),
+                                  borderSide: BorderSide(
+                                    color: Colors.white,
+                                  ),
+                                ),
+                                focusedBorder: OutlineInputBorder(
+                                  borderRadius: BorderRadius.circular(10),
+                                  borderSide: BorderSide(
+                                    color: Colors.black,
+                                  ),
+                                ),
+                                hintText: "Phone Number",
+                                hintStyle: TextStyle(color: Colors.white),
+                                border: OutlineInputBorder(
+                                  borderRadius: BorderRadius.circular(10),
+                                )),
+
+                            // InputDecoration(
+                            //   labelText: 'Phone Number',
+                            //   border: OutlineInputBorder(
+                            //     borderSide: BorderSide(),
+                            //   ),
+                            // ),
+                            initialCountryCode: 'IN',
+                            onChanged: (phone) {
+                              print(phone.completeNumber);
+                            },
+                          ),
+
+
+
+                          SizedBox(
+                            height: 30,
                           ),
                           Row(
                             mainAxisAlignment: MainAxisAlignment.spaceBetween,
